pyproject.toml
[tool.poetry]
name = "apitesting"
version = "0.1.0"
description = ""
authors = ["SeniorK0tik <ndrus24@bk.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
allure-pytest = "^2.13.2"
httpx = "^0.24.1"
jsonschema = "^4.18.0"
pydantic = "^2.0.2"
pytest = "^7.4.0"
python-dotenv = "^1.0.0"
pydantic-settings = "^2.0.1"
loguru = "^0.7.0"
pytest-asyncio = "^0.21.0"
pytest-xdist = "^3.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
